Make sure tests are reran when index.js changes
We don't (need to) require index.js in the test, because we're actually
testing the CLI via a child process. But we do want Jest to track
changes to the index.js, so tests are reran when index.js changes.

We can't however just require index.js, because it has some undesirable
side-effects (like calling process.exit). With this clever trick, the
file is not actually required, but Jest does keep track of changes in
it.